requirements.txt
Flask==0.12.1
python-baseconv==1.2.0
user-agents==1.1.0
pytest==3.0.7